start sorting out the sub-commands of `blockchain'
