+++ wallet-cli/src/command/blockchain/mod.rs
@@ -10,29 +10,13 @@ use blockchain;
 use config::{Config};
 use std::io::{Write, stdout};
 
-use exe_common::{config::{net}, network::{Peer, api::{*}}};
+use exe_common::{config::{net}, network::{api::{*}}};
 
 use command::pretty::Pretty;
 
-pub fn get_native_peer(blockchain: String, cfg: &net::Config) -> Peer {
-    for peer in cfg.peers.iter() {
-        if peer.is_native() {
-            return Peer::new(blockchain, peer.name().to_owned(), peer.peer().clone(), cfg.protocol_magic).unwrap()
-        }
-    }
+mod util;
 
-    panic!("no native peer to connect to")
-}
-
-pub fn get_http_peer(blockchain: String, cfg: &net::Config) -> Peer {
-    for peer in cfg.peers.iter() {
-        if peer.is_http() {
-            return Peer::new(blockchain, peer.name().to_owned(), peer.peer().clone(), cfg.protocol_magic).unwrap()
-        }
-    }
-
-    panic!("no native peer to connect to")
-}
+use self::util::{*, range::RangeOption};
 
 // Return the chain of block headers starting at from's next block
 // and terminating at to, unless this range represent a number
@@ -149,20 +133,6 @@ fn net_sync_faster(network: String, mut storage: Storage) {
         download_start_hash = result.next_epoch_hash.unwrap_or(result.last_header_hash);
         download_epoch_id += 1;
     }
-}
-
-fn blockchain_name_arg<'a, 'b>(index: u64) -> Arg<'a,'b> {
-    Arg::with_name("name")
-        .help("the blockchain name")
-        .index(index)
-        .required(true)
-}
-
-fn resolv_network_by_name<'a>(opts: &ArgMatches<'a>) -> Config {
-    let name = value_t!(opts.value_of("name"), String).unwrap();
-    let mut config = Config::default();
-    config.network = name;
-    config
 }
 
 pub struct Blockchain;
@@ -372,7 +342,7 @@ impl HasCommand for Blockchain {
                 let mut pack_params = PackParameters::default();
                 pack_params.delete_blobs_after_pack = ! opts.is_present("preserve-blobs");
                 if opts.is_present("range") {
-                    let range = value_t!(opts.value_of("range"), internal::RangeOption).unwrap();
+                    let range = value_t!(opts.value_of("range"), RangeOption).unwrap();
                     let from = match tag::read(&storage, &range.from) {
                         None => hex::decode(&range.from).unwrap(),
                         Some(t) => t
@@ -589,38 +559,3 @@ fn packref_fromhex(s: &String) -> PackHash {
 fn display_block(blk: &blockchain::Block) {
     println!("{}", blk.to_pretty());
 }
-
-mod internal {
-    use std::str::FromStr;
-
-    #[derive(Debug)]
-    pub struct RangeOption {
-        pub from: String,
-        pub to: Option<String>
-    }
-
-    #[derive(Debug)]
-    pub enum Error {
-        Empty,
-        InvalidRange
-    }
-
-    impl FromStr for RangeOption {
-        type Err = Error;
-
-        fn from_str(s: &str) -> Result<Self, Self::Err> {
-            if s.is_empty() { return Err(Error::Empty); }
-
-            let mut v : Vec<&str> = s.split("..").collect();
-            if v.is_empty() || v.len() > 2 { return Err(Error::InvalidRange); }
-
-            let h1 = v.pop().unwrap().to_string(); // we expect at least one
-            if let Some(h2) = v.pop().map(|v| v.to_string()) {
-                Ok(RangeOption { from: h2, to: Some(h1)})
-            } else {
-                Ok(RangeOption { from: h1, to: None})
-            }
-        }
-    }
-}
-
